+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
+- Allow to mount EmptyDir
 
 ## [1.1.0](https://github.com/arangodb/kube-arangodb/tree/master) (2020-10-14)
 - Change NumberOfCores and MemoryOverride flags to be set to true by default