+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
 - Allow to mount EmptyDir
+- Allow to specify initContainers in pods
 
 ## [1.1.0](https://github.com/arangodb/kube-arangodb/tree/master) (2020-10-14)
 - Change NumberOfCores and MemoryOverride flags to be set to true by default