+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allow to specify initContainers in pods
 - Add serviceAccount, resources and securityContext fields to ID Group
 - Allow to override Entrypoint
+- Add NodeSelector to Deployment Helm Chart
 
 ## [1.1.0](https://github.com/arangodb/kube-arangodb/tree/master) (2020-10-14)
 - Change NumberOfCores and MemoryOverride flags to be set to true by default